Silence output when running tests and rename files with internal
